Escrever para os ficheiros 3d

+++ src/Generator/calcVertices.hpp
@@ -1,11 +1,11 @@
-#include "../Utilities/utilities.hpp"
+#include "../Utilities/utilities.cpp"
 
 using namespace utilities;
 
 namespace generate{
 
-    figure createPlane(float, float);
-    figure createBox(float, float);
+    figure createPlane(float, int);
+    figure createBox(float, int);
     figure createSphere(float, int, int);
     figure createCone(float, float, int, int);
 }
